+++ archetype/simple-service/src/main/resources/archetype-resources/client/src/main/java/client/HelloWorldResource.java
@@ -1,26 +1,18 @@
 #set( $symbol_pound = '#' )
 #set( $symbol_dollar = '$' )
 #set( $symbol_escape = '\' )
-package ${package}.api;
+package ${package}.client;
 
 import ${package}.api.model.HelloWorldDto;
 import com.cerner.beadledom.jaxrs.GenericResponse;
-import com.wordnik.swagger.annotations.Api;
-import com.wordnik.swagger.annotations.ApiOperation;
-import com.wordnik.swagger.annotations.ApiResponse;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-@Api(value = "/hello", description = "Retrieve hello world data")
 @Path("/hello")
 public interface HelloWorldResource {
 
-  @ApiOperation(
-      value = "Retrieves hello world data.",
-      response = HelloWorldDto.class)
-  @ApiResponse(code = 200, message = "Success")
   @GET
   @Produces(MediaType.APPLICATION_JSON)
   GenericResponse<HelloWorldDto> getHelloWorld();
