+++ archetype/simple-service/src/main/resources/archetype-resources/service/src/main/java/service/__name__Module.java
@@ -3,14 +3,17 @@
 #set( $symbol_escape = '\' )
 package ${package}.service;
 
-import ${package}.service.resource.HelloWorldResourceImpl;
-import ${package}.api.HelloWorldResource;
+import ${package}.service.resource.HelloWorldResource;
 import com.google.inject.PrivateModule;
 
 public class ${name}Module extends PrivateModule {
 
   protected void configure() {
-    bind(HelloWorldResource.class).to(HelloWorldResourceImpl.class);
+    bind(HelloWorldResource.class);
+
+    // Note: All clients that will be used should be installed in this module to avoid registering
+    // client JAX-RS resources with the server. The Correlation ID header name for clients could
+    // also be configured/overridden here.
 
     expose(HelloWorldResource.class);
   }
