+++ pom.xml
@@ -89,7 +89,7 @@
     <scala.version>2.11.12</scala.version>
     <scala.binary.version>2.11</scala.binary.version>
     <governator.version>1.17.4</governator.version>
-    <jackson.version>2.9.2</jackson.version>
+    <jackson.version>2.9.5</jackson.version>
     <resteasy.version>3.1.4.Final</resteasy.version>
     <spotbugs.version>3.1.1</spotbugs.version>
     <stagemonitor.version>0.22.0</stagemonitor.version>
